SPF - changed from blacklist to blocklist
Both terms refer to a list of denied items, but "blocklist" is the modern and preferred term for inclusivity. While "blacklist" is still widely understood and used, "blocklist" is being adopted by many organizations to replace the exclusionary term.

We try to use blocklist in both UI and our documentation.